backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Audio processing libraries
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.4
pydub==0.25.1

# Speech-to-text with OpenAI Whisper API
openai==1.6.1

# Text comparison
python-Levenshtein==0.23.0
